Move discount releated fixtures from global fixture file
