+++ saleor/discount/tests/fixtures/__init__.py
@@ -1,0 +1,3 @@
+from .voucher import *  # noqa: F403
+from .promotion import *  # noqa: F403
+from .manual_discount import *  # noqa: F403
