+++ saleor/tests/fixtures.py
@@ -1868,133 +1868,6 @@ def image_list():
 
 
 @pytest.fixture
-def voucher_without_channel(db):
-    voucher = Voucher.objects.create()
-    VoucherCode.objects.create(code="mirumee", voucher=voucher)
-    return voucher
-
-
-@pytest.fixture
-def voucher(voucher_without_channel, channel_USD):
-    VoucherChannelListing.objects.create(
-        voucher=voucher_without_channel,
-        channel=channel_USD,
-        discount=Money(20, channel_USD.currency_code),
-    )
-    return voucher_without_channel
-
-
-@pytest.fixture
-def voucher_with_many_codes(voucher):
-    VoucherCode.objects.bulk_create(
-        [
-            VoucherCode(code="Multi1", voucher=voucher),
-            VoucherCode(code="Multi2", voucher=voucher),
-            VoucherCode(code="Multi3", voucher=voucher),
-            VoucherCode(code="Multi4", voucher=voucher),
-        ]
-    )
-    return voucher
-
-
-@pytest.fixture
-def voucher_with_many_channels(voucher, channel_PLN):
-    VoucherChannelListing.objects.create(
-        voucher=voucher,
-        channel=channel_PLN,
-        discount=Money(80, channel_PLN.currency_code),
-    )
-    return voucher
-
-
-@pytest.fixture
-def voucher_percentage(channel_USD):
-    voucher = Voucher.objects.create(
-        discount_value_type=DiscountValueType.PERCENTAGE,
-    )
-    VoucherCode.objects.create(code="saleor", voucher=voucher)
-    VoucherChannelListing.objects.create(
-        voucher=voucher,
-        channel=channel_USD,
-        discount_value=10,
-        currency=channel_USD.currency_code,
-    )
-    return voucher
-
-
-@pytest.fixture
-def voucher_specific_product_type(voucher_percentage, product):
-    voucher_percentage.products.add(product)
-    voucher_percentage.type = VoucherType.SPECIFIC_PRODUCT
-    voucher_percentage.save()
-    return voucher_percentage
-
-
-@pytest.fixture
-def voucher_with_high_min_spent_amount(channel_USD):
-    voucher = Voucher.objects.create()
-    VoucherCode.objects.create(code="mirumee", voucher=voucher)
-    VoucherChannelListing.objects.create(
-        voucher=voucher,
-        channel=channel_USD,
-        discount=Money(10, channel_USD.currency_code),
-        min_spent_amount=1_000_000,
-    )
-    return voucher
-
-
-@pytest.fixture
-def voucher_shipping_type(channel_USD):
-    voucher = Voucher.objects.create(type=VoucherType.SHIPPING, countries="IS")
-    VoucherCode.objects.create(code="mirumee", voucher=voucher)
-    VoucherChannelListing.objects.create(
-        voucher=voucher,
-        channel=channel_USD,
-        discount=Money(10, channel_USD.currency_code),
-    )
-    return voucher
-
-
-@pytest.fixture
-def voucher_free_shipping(voucher_percentage, channel_USD):
-    voucher_percentage.type = VoucherType.SHIPPING
-    voucher_percentage.name = "Free shipping"
-    voucher_percentage.save()
-    voucher_percentage.channel_listings.filter(channel=channel_USD).update(
-        discount_value=100
-    )
-    return voucher_percentage
-
-
-@pytest.fixture
-def voucher_customer(voucher, customer_user):
-    email = customer_user.email
-    code = voucher.codes.first()
-    return VoucherCustomer.objects.create(voucher_code=code, customer_email=email)
-
-
-@pytest.fixture
-def voucher_multiple_use(voucher_with_many_codes):
-    voucher = voucher_with_many_codes
-    voucher.usage_limit = 3
-    voucher.save(update_fields=["usage_limit"])
-    codes = voucher.codes.all()
-    for code in codes:
-        code.used = 1
-    VoucherCode.objects.bulk_update(codes, ["used"])
-    voucher.refresh_from_db()
-    return voucher
-
-
-@pytest.fixture
-def voucher_single_use(voucher_with_many_codes):
-    voucher = voucher_with_many_codes
-    voucher.single_use = True
-    voucher.save(update_fields=["single_use"])
-    return voucher
-
-
-@pytest.fixture
 def checkout_line_with_reservation_in_many_stocks(
     customer_user, variant_with_many_stocks, checkout
 ):
@@ -2237,345 +2110,6 @@ def dummy_address_data(address):
 def dummy_webhook_app_payment_data(dummy_payment_data, payment_app):
     dummy_payment_data.gateway = to_payment_app_id(payment_app, "credit-card")
     return dummy_payment_data
-
-
-@pytest.fixture
-def catalogue_promotion(channel_USD, product, collection):
-    promotion = Promotion.objects.create(
-        name="Promotion",
-        type=PromotionType.CATALOGUE,
-        description=dummy_editorjs("Test description."),
-        end_date=timezone.now() + datetime.timedelta(days=30),
-    )
-    rules = PromotionRule.objects.bulk_create(
-        [
-            PromotionRule(
-                name="Percentage promotion rule",
-                promotion=promotion,
-                description=dummy_editorjs(
-                    "Test description for percentage promotion rule."
-                ),
-                catalogue_predicate={
-                    "productPredicate": {
-                        "ids": [graphene.Node.to_global_id("Product", product.id)]
-                    }
-                },
-                reward_value_type=RewardValueType.PERCENTAGE,
-                reward_value=Decimal("10"),
-            ),
-            PromotionRule(
-                name="Fixed promotion rule",
-                promotion=promotion,
-                description=dummy_editorjs(
-                    "Test description for fixes promotion rule."
-                ),
-                catalogue_predicate={
-                    "collectionPredicate": {
-                        "ids": [graphene.Node.to_global_id("Collection", collection.id)]
-                    }
-                },
-                reward_value_type=RewardValueType.FIXED,
-                reward_value=Decimal("5"),
-            ),
-        ]
-    )
-    for rule in rules:
-        rule.channels.add(channel_USD)
-    fetch_variants_for_promotion_rules(promotion.rules.all())
-    return promotion
-
-
-@pytest.fixture
-def catalogue_promotion_without_rules(db):
-    promotion = Promotion.objects.create(
-        name="Promotion",
-        description=dummy_editorjs("Test description."),
-        end_date=timezone.now() + datetime.timedelta(days=30),
-        type=PromotionType.CATALOGUE,
-    )
-    return promotion
-
-
-@pytest.fixture
-def order_promotion_without_rules(db):
-    promotion = Promotion.objects.create(
-        name="Promotion",
-        description=dummy_editorjs("Test description."),
-        end_date=timezone.now() + datetime.timedelta(days=30),
-        type=PromotionType.ORDER,
-    )
-    return promotion
-
-
-@pytest.fixture
-def catalogue_promotion_with_single_rule(catalogue_predicate, channel_USD):
-    promotion = Promotion.objects.create(
-        name="Promotion with single rule", type=PromotionType.CATALOGUE
-    )
-    rule = PromotionRule.objects.create(
-        name="Sale rule",
-        promotion=promotion,
-        catalogue_predicate=catalogue_predicate,
-        reward_value_type=RewardValueType.FIXED,
-        reward_value=Decimal(5),
-    )
-    rule.channels.add(channel_USD)
-    return promotion
-
-
-@pytest.fixture
-def order_promotion_with_rule(channel_USD):
-    promotion = Promotion.objects.create(
-        name="Promotion with order rule", type=PromotionType.ORDER
-    )
-    rule = PromotionRule.objects.create(
-        name="Promotion rule",
-        promotion=promotion,
-        order_predicate={
-            "discountedObjectPredicate": {"baseSubtotalPrice": {"range": {"gte": 100}}}
-        },
-        reward_value_type=RewardValueType.FIXED,
-        reward_value=Decimal(5),
-        reward_type=RewardType.SUBTOTAL_DISCOUNT,
-    )
-    rule.channels.add(channel_USD)
-    return promotion
-
-
-@pytest.fixture
-def promotion_list(channel_USD, product, collection):
-    collection.products.add(product)
-    promotions = Promotion.objects.bulk_create(
-        [
-            Promotion(
-                name="Promotion 1",
-                type=PromotionType.CATALOGUE,
-                description=dummy_editorjs("Promotion 1 description."),
-                start_date=timezone.now() + datetime.timedelta(days=1),
-                end_date=timezone.now() + datetime.timedelta(days=10),
-            ),
-            Promotion(
-                name="Promotion 2",
-                type=PromotionType.CATALOGUE,
-                description=dummy_editorjs("Promotion 2 description."),
-                start_date=timezone.now() + datetime.timedelta(days=5),
-                end_date=timezone.now() + datetime.timedelta(days=20),
-            ),
-            Promotion(
-                name="Promotion 3",
-                type=PromotionType.CATALOGUE,
-                description=dummy_editorjs("TePromotion 3 description."),
-                start_date=timezone.now() + datetime.timedelta(days=15),
-                end_date=timezone.now() + datetime.timedelta(days=30),
-            ),
-        ]
-    )
-    rules = PromotionRule.objects.bulk_create(
-        [
-            PromotionRule(
-                name="Promotion 1 percentage rule",
-                promotion=promotions[0],
-                description=dummy_editorjs(
-                    "Test description for promotion 1 percentage rule."
-                ),
-                catalogue_predicate={
-                    "productPredicate": {
-                        "ids": [graphene.Node.to_global_id("Product", product.id)]
-                    }
-                },
-                reward_value_type=RewardValueType.PERCENTAGE,
-                reward_value=Decimal("10"),
-            ),
-            PromotionRule(
-                name="Promotion 1 fixed rule",
-                promotion=promotions[0],
-                description=dummy_editorjs(
-                    "Test description for promotion 1 fixed rule."
-                ),
-                catalogue_predicate={
-                    "collectionPredicate": {
-                        "ids": [graphene.Node.to_global_id("Collection", collection.id)]
-                    }
-                },
-                reward_value_type=RewardValueType.FIXED,
-                reward_value=Decimal("5"),
-            ),
-            PromotionRule(
-                name="Promotion 2 percentage rule",
-                promotion=promotions[1],
-                description=dummy_editorjs(
-                    "Test description for promotion 2 percentage rule."
-                ),
-                catalogue_predicate={
-                    "productPredicate": {
-                        "ids": [graphene.Node.to_global_id("Product", product.id)]
-                    }
-                },
-                reward_value_type=RewardValueType.PERCENTAGE,
-                reward_value=Decimal("10"),
-            ),
-            PromotionRule(
-                name="Promotion 3 fixed rule",
-                promotion=promotions[2],
-                description=dummy_editorjs(
-                    "Test description for promotion 3 fixed rule."
-                ),
-                catalogue_predicate={
-                    "collectionPredicate": {
-                        "ids": [graphene.Node.to_global_id("Collection", collection.id)]
-                    }
-                },
-                reward_value_type=RewardValueType.FIXED,
-                reward_value=Decimal("5"),
-            ),
-        ]
-    )
-    for rule in rules:
-        rule.channels.add(channel_USD)
-    fetch_variants_for_promotion_rules(PromotionRule.objects.all())
-    return promotions
-
-
-@pytest.fixture
-def promotion_rule(channel_USD, catalogue_promotion, product):
-    rule = PromotionRule.objects.create(
-        name="Promotion rule name",
-        promotion=catalogue_promotion,
-        description=dummy_editorjs("Test description for percentage promotion rule."),
-        catalogue_predicate={
-            "productPredicate": {
-                "ids": [graphene.Node.to_global_id("Product", product.id)]
-            }
-        },
-        reward_value_type=RewardValueType.PERCENTAGE,
-        reward_value=Decimal("25"),
-    )
-    rule.channels.add(channel_USD)
-    return rule
-
-
-@pytest.fixture
-def order_promotion_rule(channel_USD, order_promotion_without_rules):
-    rule = PromotionRule.objects.create(
-        name="Order promotion rule",
-        promotion=order_promotion_without_rules,
-        order_predicate={
-            "discountedObjectPredicate": {"baseSubtotalPrice": {"range": {"gte": 20}}}
-        },
-        reward_value_type=RewardValueType.PERCENTAGE,
-        reward_value=Decimal("25"),
-        reward_type=RewardType.SUBTOTAL_DISCOUNT,
-    )
-    rule.channels.add(channel_USD)
-    return rule
-
-
-@pytest.fixture
-def gift_promotion_rule(channel_USD, order_promotion_without_rules, product_list):
-    rule = PromotionRule.objects.create(
-        name="Order promotion rule",
-        promotion=order_promotion_without_rules,
-        order_predicate={
-            "discountedObjectPredicate": {"baseSubtotalPrice": {"range": {"gte": 20}}}
-        },
-        reward_type=RewardType.GIFT,
-    )
-    rule.channels.add(channel_USD)
-    rule.gifts.set([product.variants.first() for product in product_list[:2]])
-    return rule
-
-
-@pytest.fixture
-def rule_info(
-    promotion_rule,
-    promotion_translation_fr,
-    promotion_rule_translation_fr,
-    variant,
-    channel_USD,
-):
-    variant_channel_listing = variant.channel_listings.get(channel_id=channel_USD.id)
-    listing_promotion_rule = variant_channel_listing.variantlistingpromotionrule.create(
-        promotion_rule=promotion_rule,
-        discount_amount=Decimal("10"),
-        currency=channel_USD.currency_code,
-    )
-    return VariantPromotionRuleInfo(
-        rule=promotion_rule,
-        promotion=promotion_rule.promotion,
-        variant_listing_promotion_rule=listing_promotion_rule,
-        promotion_translation=promotion_translation_fr,
-        rule_translation=promotion_rule_translation_fr,
-    )
-
-
-@pytest.fixture
-def catalogue_predicate(product, category, collection, variant):
-    collection_id = graphene.Node.to_global_id("Collection", collection.id)
-    category_id = graphene.Node.to_global_id("Category", category.id)
-    product_id = graphene.Node.to_global_id("Product", product.id)
-    variant_id = graphene.Node.to_global_id("ProductVariant", variant.id)
-    return {
-        "OR": [
-            {"collectionPredicate": {"ids": [collection_id]}},
-            {"categoryPredicate": {"ids": [category_id]}},
-            {"productPredicate": {"ids": [product_id]}},
-            {"variantPredicate": {"ids": [variant_id]}},
-        ]
-    }
-
-
-@pytest.fixture
-def promotion_converted_from_sale(catalogue_predicate, channel_USD):
-    promotion = Promotion.objects.create(name="Sale", type=PromotionType.CATALOGUE)
-    promotion.assign_old_sale_id()
-
-    rule = PromotionRule.objects.create(
-        name="Sale rule",
-        promotion=promotion,
-        catalogue_predicate=catalogue_predicate,
-        reward_value_type=RewardValueType.FIXED,
-        reward_value=Decimal(5),
-        old_channel_listing_id=PromotionRule.get_old_channel_listing_ids(1)[0][0],
-    )
-    rule.channels.add(channel_USD)
-    fetch_variants_for_promotion_rules(promotion.rules.all())
-    return promotion
-
-
-@pytest.fixture
-def promotion_converted_from_sale_with_many_channels(
-    promotion_converted_from_sale, catalogue_predicate, channel_PLN
-):
-    promotion = promotion_converted_from_sale
-    rule = PromotionRule.objects.create(
-        name="Sale rule 2",
-        promotion=promotion,
-        catalogue_predicate=catalogue_predicate,
-        reward_value_type=RewardValueType.FIXED,
-        reward_value=Decimal(5),
-        old_channel_listing_id=PromotionRule.get_old_channel_listing_ids(1)[0][0],
-    )
-    rule.channels.add(channel_PLN)
-    fetch_variants_for_promotion_rules(promotion.rules.all())
-    return promotion
-
-
-@pytest.fixture
-def promotion_converted_from_sale_with_empty_predicate(channel_USD):
-    promotion = Promotion.objects.create(
-        name="Sale with empty predicate", type=PromotionType.CATALOGUE
-    )
-    promotion.assign_old_sale_id()
-    rule = PromotionRule.objects.create(
-        name="Sale with empty predicate rule",
-        promotion=promotion,
-        catalogue_predicate={},
-        reward_value_type=RewardValueType.FIXED,
-        reward_value=Decimal(5),
-        old_channel_listing_id=PromotionRule.get_old_channel_listing_ids(1)[0][0],
-    )
-    rule.channels.add(channel_USD)
-    return promotion
 
 
 @pytest.fixture
